refactor(db): create `database` module
